refactor: change casing of controller files for consistency
