+++ api/server/controllers/auth/LogoutController.js
