+++ api/server/index.js
@@ -6,7 +6,7 @@ const indexSync = require('../lib/db/indexSync');
 const path = require('path');
 const cors = require('cors');
 const routes = require('./routes');
-const errorController = require('./controllers/error.controller');
+const errorController = require('./controllers/ErrorController');
 const passport = require('passport');
 const port = process.env.PORT || 3080;
 const host = process.env.HOST || 'localhost';
